refactor(emails): lift DetailRow into EmailLayout as a shared export
The verification email's local DetailRow (plus its label/value column
styles) becomes a shared EmailLayout export so the registration
notification can reuse the same fact-table row. EventVerificationEmail
now imports it; its rendered output is unchanged.

Refs #588
